requirements.txt
pyobjc-framework-FSEvents